2051: move WriteCsv to bayern. add config output path
